// modules/core/src/types/Tag.ts
export enum Tag {
  A = 'a',
  ABBR = 'abbr',
  ADDRESS = 'address',
  AREA = 'area',
  ARTICLE = 'article',
  ASIDE = 'aside',
  AUDIO = 'audio',
  B = 'b',
  BASE = 'base',
  BDO = 'bdo',
  BLOCKQUOTE = 'blockquote',
  BODY = 'body',
  BR = 'br',
  BUTTON = 'button',
  CANVAS = 'canvas',
  CAPTION = 'caption',
  CITE = 'cite',
  CODE = 'code',
  COL = 'col',
  COLGROUP = 'colgroup',
  COMMAND = 'command',
  DATALIST = 'datalist',
  DD = 'dd',
  DEL = 'del',
  DETAILS = 'details',
  DFN = 'dfn',
  DIV = 'div',
  DL = 'dl',
  DT = 'dt',
  EM = 'em',
  EMBED = 'embed',
  FIELDSET = 'fieldset',
  FIGCAPTION = 'figcaption',
  FIGURE = 'figure',
  FOOTER = 'footer',
  FORM = 'form',
  H1 = 'h1',
  H2 = 'h2',
  H3 = 'h3',
  H4 = 'h4',
  H5 = 'h5',
  H6 = 'h6',
  HEAD = 'head',
  HEADER = 'header',
  HR = 'hr',
  HTML = 'html',
  I = 'i',
  IFRAME = 'iframe',
  IMG = 'img',
  INPUT = 'input',
  INS = 'ins',
  KBD = 'kbd',
  LABEL = 'label',
  LEGEND = 'legend',
  LI = 'li',
  LINK = 'link',
  MAP = 'map',
  MARK = 'mark',
  MENU = 'menu',
  META = 'meta',
  METER = 'meter',
  NAV = 'nav',
  NOSCRIPT = 'noscript',
  OBJECT = 'object',
  OL = 'ol',
  OPTGROUP = 'optgroup',
  OPTION = 'option',
  OUTPUT = 'output',
  P = 'p',
  PARAM = 'param',
  PRE = 'pre',
  PROGRESS = 'progress',
  Q = 'q',
  RP = 'rp',
  RT = 'rt',
  RUBY = 'ruby',
  S = 's',
  SAMP = 'samp',
  SCRIPT = 'script',
  SECTION = 'section',
  SELECT = 'select',
  SMALL = 'small',
  SOURCE = 'source',
  SPAN = 'span',
  STRONG = 'strong',
  STYLE = 'style',
  SUB = 'sub',
  SUP = 'sup',
  TABLE = 'table',
  TBODY = 'tbody',
  TD = 'td',
  TEXTAREA = 'textarea',
  TFOOT = 'tfoot',
  TH = 'th',
  THEAD = 'thead',
  TIME = 'time',
  TITLE = 'title',
  TR = 'tr',
  TRACK = 'track',
  U = 'u',
  UL = 'ul',
  VAR = 'var',
  VIDEO = 'video',
  WBR = 'wbr'
}